feat: extract shared NewsletterView with better error reporting
- Extract newsletter signup into reusable NewsletterView component
- Add friendly error messages for 429, 409, and network failures
- Add source field to subscriber API requests
