+++ src/Ivy.Tendril/Apps/HelpApp.cs
@@ -1,5 +1,4 @@
-using System.Net.Http.Json;
-using Ivy.Tendril.Helpers;
+using Ivy.Tendril.Apps.Views;
 using Ivy.Tendril.Services;
 
 namespace Ivy.Tendril.Apps;
@@ -10,12 +9,6 @@ public class HelpApp : ViewBase
     public override object Build()
     {
         var client = UseService<IClientProvider>();
-        var httpClientFactory = UseService<IHttpClientFactory>();
-        var telemetry = UseService<ITelemetryService>();
-        var email = UseState("");
-        var subscribed = UseState(false);
-        var error = UseState<string?>(null);
-        var isLoading = UseState(false);
 
         return Layout.TopCenter()
                | (Layout.Vertical().Margin(0, 20).Width(150)
@@ -40,50 +33,7 @@ public class HelpApp : ViewBase
                   | new Separator()
                   | Text.H2("Newsletter")
                   | Text.Muted("Be the first to know when we have a new release!")
-                  | (subscribed.Value
-                      ? Text.Success("Subscribed!")
-                      : (Layout.Horizontal()
-                         | email.ToTextInput("you@example.com")
-                         | new Button("Subscribe")
-                             .Primary()
-                             .Disabled(!InputSanitizer.IsValidEmail(email.Value))
-                             .Loading(isLoading.Value)
-                             .OnClick(Subscribe)))
-                  | (error.Value != null ? Text.Danger(error.Value) : null)
+                  | new NewsletterView()
                );
-
-        async ValueTask Subscribe(Event<Button> e)
-        {
-            if (!InputSanitizer.IsValidEmail(email.Value))
-            {
-                error.Value = "Please enter a valid email address.";
-                return;
-            }
-
-            isLoading.Value = true;
-            try
-            {
-                using var http = httpClientFactory.CreateClient();
-                var response = await http.PostAsJsonAsync("https://tendril-api.ivy.app/subscribers", new
-                {
-                    email = email.Value,
-                    anonymousId = telemetry.AnonymousId
-                });
-                response.EnsureSuccessStatusCode();
-                subscribed.Value = true;
-                error.Value = null;
-            }
-            catch (Exception ex)
-            {
-                error.Value = ex.Message;
-            }
-            finally
-            {
-                isLoading.Value = false;
-            }
-        }
-
     }
 }
-
-
